Import database client while in use
As we support more and more databases, vectordb-bench
depends on more and more database clients, which causes some
dependency problems. Especially when we only cares for one or two
databases.

This PR removes the dependency of all database clients. We are still
able to browse the standard results with no database clients installed.

And we can install the database clients when we need to do benchmarks on
the sepcific database.

1. The defatul vectordb-bench now only contains PyMilvus

        pip install vectordb-bench # pymilvus only

2. We can still install all database clients
        pip install vectordb-bench[all]

3. Or some other database client
        pip install vectordb-bench[pinecone]

Signed-off-by: yangxuan <xuan.yang@zilliz.com>